ARM64: dts: Add initial DT support for simtemp dev
This commit introduces a new devicetree overlay for the 'simtemp'
simulated temperature device. The overlay registers the device,
allowing the kernel to recognize and interface with it.

Also included are the necessary build scripts to compile the
overlay and a skeleton kernel mode driver that binds to the
'simtemp' device.

To facilitate testing and development, the project's README has
been updated to include instructions on deploying the DTBO in a
QEMU environment for the ARM64 Raspberry Pi architecture.