nxp_simtemp: Enable driver to load w/o DT entry
The current driver depends on a corresponding device tree entry
to be present for a successful probe. However, on some systems,
ACPI handled, this device may not be described in the DTB.

To ensure the driver always loads when available, add a fallback
mechanism to create a platform device if no matching device is
found via the DT.

This ensures the driver is always available for systems where
ACPI handles device discovery.